Added plus button to menu for adding ingredients

+++ app/src/main/java/com/example/capstone18/capstone/Ingredient.java
@@ -28,7 +28,7 @@ public class Ingredient {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
